requirements.txt
# Core dependencies
python-dotenv>=1.0.0
oracledb>=1.4.0
langchain>=0.0.200
langchain-community>=0.0.10
sentence-transformers>=2.2.2
chromadb>=0.4.0
openai>=0.27.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
